L6 G3b: firmware GSP por familia y primer tramo de la ruta FMC de Blackwell
Firmware chip-aware: tabla por familia en gsp_fw.c — Blackwell pide
gb205/{bootloader,fmc,gsp}, Ampere ga102/{bootloader,booter_load,
booter_unload,gsp} — y gsp_fw_load_all(chip) exige todos los blobs de su
juego. Antes cargaba los dos y duplicaba 60,6 MiB de ucode, porque en
linux-firmware el gsp de gb205 es un symlink al de ga102; con eso el heap se
agotaba en el staging GEM. Empaqueta los blobs ga102 que faltaban.

Bring-up bifurcado por familia: Ampere sigue con el ACR de SEC2, Blackwell
entra en la ruta FMC. El ACR es de Ampere y en GB205 daba mbox0=0xbadf4100
(el falcon ni ejecuta) porque GB20x arranca por GSP-FMC/FSP.

fmc_lx.c: valida el ELF firmado (cabecera fija, secciones, CRC32) y extrae
hash/signature/publickey/image, comprueba los tamaños COT de gb202 (48/97/96)
y lee el estado del FSP. Verificado en HW: el ELF cuadra y el FSP responde
secure boot=0xff con el canal EMEM ocioso. No escribe MMIO: el mensaje COT
necesita antes WPR meta + radix3 + libos boot args.

Referencia upstream (no está en el árbol 6.6 pinneado): nvkm/subdev/fsp/
{gh100,gb202}.c y nvkm/subdev/gsp/gh100.c.

De paso, la VRAM se truncaba a 0: 12 GiB es 3*2^32 exactos y el tipo era
unsigned de 32 bits. Toda la cadena pasa a 64 bits.

+++ kernel/src/lxdde/gpu.rs
@@ -13,7 +13,7 @@ unsafe extern "C" {
         x: *const f32,
         y: *mut f32,
     ) -> i32;
-    fn lx_nouveau_vram_total() -> u32;
+    fn lx_nouveau_vram_total() -> u64;
     fn lx_nouveau_set_boot0(boot0: u32, device_id: u32);
 }
 
@@ -42,7 +42,7 @@ pub fn gsp_phase() -> &'static str {
 }
 
 pub fn vram_total() -> u64 {
-    unsafe { lx_nouveau_vram_total() as u64 }
+    unsafe { lx_nouveau_vram_total() }
 }
 
 pub fn submit_saxpy(a: f32, x: &[f32], y: &mut [f32]) -> Result<bool, ()> {
